Core: Cleanup of TrackingStruct tests
1) Split the TrackingBuilder tests to a separate suite
2) Replaced setters by ordinal usage with constructors calls, except in
   tests that deliberately exercise these setters
3) In TestTrackingStruct replaced the usage of the builder with
   constructor calls
